# ref/openapi_v2.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Manus OpenAPI v2",
        "description": "API for integrating Manus into your workflow. All responses are wrapped with {\"ok\": true, \"request_id\": \"...\", ...} for success and {\"ok\": false, \"request_id\": \"...\", \"error\": {\"code\": \"...\", \"message\": \"...\"}} for errors.",
        "version": "2.0.0"
    },
    "servers": [
        {
            "url": "https://api.manus.ai"
        }
    ],
    "security": [
        {
            "ApiKeyAuth": []
        }
    ],
    "paths": {
        "/v2/task.create": {
            "post": {
                "summary": "CreateTask",
                "description": "Creates a new task. The task runs asynchronously. Poll for progress with [task.listMessages](https://open.manus.ai/docs/v2/task.listMessages), send follow-ups with [task.sendMessage](https://open.manus.ai/docs/v2/task.sendMessage). See the [Task Lifecycle](https://open.manus.ai/docs/v2/task-lifecycle) guide for the complete flow.",
                "operationId": "openapi.v2.OpenapiV2Service.CreateTask",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "message": {
                                        "$ref": "#/components/schemas/Message",
                                        "description": "The message to start the task with. Contains the prompt text, optional file attachments, and connector/skill configuration."
                                    },
                                    "project_id": {
                                        "type": "string",
                                        "description": "Project ID to associate this task with. The project's instruction will be automatically applied. Use [project.list](https://open.manus.ai/docs/v2/project.list) to get available project IDs."
                                    },
                                    "locale": {
                                        "type": "string",
                                        "description": "Locale for the task output language (e.g., \"en\", \"zh-CN\", \"ja\"). Defaults to the user's account locale setting."
                                    },
                                    "interactive_mode": {
                                        "type": "boolean",
                                        "description": "When enabled, the agent may pause and ask follow-up questions if the input is insufficient. When disabled (default), the agent proceeds with best-effort execution without asking.",
                                        "default": false
                                    },
                                    "hide_in_task_list": {
                                        "type": "boolean",
                                        "description": "When true, the task will not appear in the Manus webapp task list. The task is still accessible via the task_url in the response. Useful for automated/background tasks.",
                                        "default": false
                                    },
                                    "share_visibility": {
                                        "type": "string",
                                        "enum": [
                                            "private",
                                            "team",
                                            "public"
                                        ],
                                        "description": "Controls who can view the task. \"private\" (default) — only the task creator can view. \"team\" — all team members can view. \"public\" — anyone with the share_url can view without authentication.",
                                        "default": "private"
                                    },
                                    "agent_profile": {
                                        "type": "string",
                                        "enum": [
                                            "manus-1.6",
                                            "manus-1.6-lite",
                                            "manus-1.6-max"
                                        ],
                                        "description": "Agent profile to use for the task. \"manus-1.6\" (default) — standard capability. \"manus-1.6-lite\" — lightweight, faster responses. \"manus-1.6-max\" — maximum capability. Free personal accounts are downgraded to `manus-1.6-lite` regardless of the requested value.",
                                        "default": "manus-1.6"
                                    },
                                    "title": {
                                        "type": "string",
                                        "description": "Custom title for the task. If not provided, a title will be auto-generated based on the input message."
                                    },
                                    "structured_output_schema": {
                                        "type": "object",
                                        "description": "JSON Schema for structured output extraction. When provided, the agent runs normally, then a post-processing step extracts a result conforming to your schema. The schema must follow the [Structured Output](https://open.manus.ai/docs/v2/structured-output) subset of JSON Schema: all object properties must be listed in `required`, and `additionalProperties` must be `false`. See the [Structured Output](https://open.manus.ai/docs/v2/structured-output) guide."
                                    }
                                },
                                "required": [
                                    "message"
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Task created successfully.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "ok": {
                                            "type": "boolean",
                                            "example": true,
                                            "description": "Whether the request was successful."
                                        },
                                        "request_id": {
                                            "type": "string",
                                            "description": "Unique identifier for this API request, useful for debugging."
                                        },
                                        "task_id": {
                                            "type": "string",
                                            "description": "Unique identifier for the created task. Use this to track, update, or delete the task."
                                        },
                                        "task_title": {
                                            "type": "string",
                                            "description": "Title for the task. Returns the custom title if one was provided in the request, otherwise an auto-generated title based on the input message."
                                        },
                                        "task_url": {
                                            "type": "string",
                                            "description": "URL to view the task in the Manus webapp (e.g., https://manus.im/app/{task_id})."
                                        },
                                        "share_url": {
                                            "type": "string",
                                            "description": "Publicly accessible URL for sharing. Only present when share_visibility is not \"private\"."
                                        },
                                        "share_visibility": {
                                            "type": "string",
                                            "enum": [
                                                "private",
                                                "team",
                                                "public"
                                            ],
                                            "description": "The actual visibility state of the task."
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "4XX": {
                        "description": "Error response.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [],
                "parameters": [
                    {
                        "name": "x-manus-api-key",
                        "in": "header",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "description": "API key for direct authentication. Provide either this or `Authorization`, not both. See [Authentication](https://open.manus.ai/docs/v2/authentication)."
                    },
                    {
                        "name": "Authorization",
                        "in": "header",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "Bearer {access_token}"
                        },
                        "description": "OAuth2 access token in `Bearer {token}` format. Provide either this or `x-manus-api-key`, not both. See the [Open App](https://open.manus.ai/docs/v2/open-app) guide."
                    }
                ]
            }
        },
...
        "/v2/task.list": {
            "get": {
                "summary": "ListTasks",
                "description": "Lists tasks with optional filtering and cursor-based pagination. Use `scope` to filter by task type, e.g. `agent_subtask` with `agent_id` for agent subtasks. See [task.detail](https://open.manus.ai/docs/v2/task.detail) for full details on a specific task.",
                "operationId": "openapi.v2.OpenapiV2Service.ListTasks",
                "parameters": [
                    {
                        "name": "x-manus-api-key",
                        "in": "header",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "description": "API key for direct authentication. Provide either this or `Authorization`, not both. See [Authentication](https://open.manus.ai/docs/v2/authentication)."
                    },
                    {
                        "name": "Authorization",
                        "in": "header",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "Bearer {access_token}"
                        },
                        "description": "OAuth2 access token in `Bearer {token}` format. Provide either this or `x-manus-api-key`, not both. See the [Open App](https://open.manus.ai/docs/v2/open-app) guide."
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "description": "Number of tasks to return per page. Default: 20, Max: 100."
                        }
                    },
                    {
                        "name": "cursor",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "description": "Pagination cursor from the previous response's next_cursor field. Omit for the first page."
                        }
                    },
                    {
                        "name": "order",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "asc",
                                "desc"
                            ],
                            "description": "Sort direction by creation time. \"desc\" (default) returns newest first, \"asc\" returns oldest first."
                        }
                    },
                    {
                        "name": "scope",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "all",
                                "agent_subtask",
                                "project",
                                "standard"
                            ],
                            "description": "Filter by task type. \"all\" (default) returns all tasks. \"standard\" returns regular tasks. \"project\" returns tasks within a project (requires project_id). \"agent_subtask\" returns subtasks created by an agent (requires agent_id)."
                        }
                    },
                    {
                        "name": "agent_id",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "description": "Filter tasks by agent. Required when scope=\"agent_subtask\". Supports the shortcut `agent-default` for the IM agent. Use [agent.list](https://open.manus.ai/docs/v2/agent.list) to get available agent IDs."
                        }
                    },
                    {
                        "name": "project_id",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "description": "Filter tasks by project. Required when scope=\"project\". Use [project.list](https://open.manus.ai/docs/v2/project.list) to get available project IDs."
                        }
                    },
                    {
                        "name": "oauth_client_id",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "description": "Filter tasks by the Open App (OAuth client) that created them. Mutually exclusive with api_key_id — passing both returns an invalid_argument error. OAuth tokens with only the create_task scope are already limited to their own app's tasks and may only pass their own client ID; passing another app's client ID returns a permission_denied error. IDs that don't match any of the user's tasks return an empty list."
                        }
                    },
                    {
                        "name": "api_key_id",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "description": "Filter tasks by the API key that created them. Mutually exclusive with oauth_client_id — passing both returns an invalid_argument error. Not available to OAuth tokens with only the create_task scope (returns a permission_denied error). IDs that don't match any of the user's tasks return an empty list."
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tasks retrieved successfully.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "ok": {
                                            "type": "boolean",
                                            "example": true,
                                            "description": "Whether the request was successful."
                                        },
                                        "request_id": {
                                            "type": "string",
                                            "description": "Unique identifier for this API request."
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Task"
                                            },
                                            "description": "Array of task objects matching the filter criteria."
                                        },
                                        "has_more": {
                                            "type": "boolean",
                                            "description": "Whether there are more tasks beyond this page. If true, use next_cursor to fetch the next page."
                                        },
                                        "next_cursor": {
                                            "type": "string",
                                            "description": "Cursor to pass as the cursor parameter for the next page. Only present when has_more is true."
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "4XX": {
                        "description": "Error response.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": []
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "ApiKeyAuth": {
                "type": "apiKey",
                "in": "header",
                "name": "x-manus-api-key"
            }
        },
        "schemas": {
            "Message": {
                "type": "object",
                "description": "A message to send to a task.",
                "properties": {
                    "content": {
                        "description": "String (plain text) or array of ContentPart objects. User text is limited to approximately 5,000 estimated tokens per request; the exact character count varies by language and content. For an array, the limit applies to the combined text of all text parts; file parts do not count. Input over the limit returns InvalidArgument (HTTP 400).",
                        "oneOf": [
                            {
                                "type": "array",
                                "items": {
                                    "$ref": "#/components/schemas/ContentPart"
                                }
                            },
                            {
                                "type": "string"
                            }
                        ]
                    },
                    "connectors": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "description": "List of connector IDs to enable for this task. Resolution order when omitted: 1) if the task belongs to a project, the project's default connectors are used; 2) otherwise, the user's default enabled connectors are used. Use [connector.list](https://open.manus.ai/docs/v2/connector.list) to get the connectors installed in your account."
                    },
                    "enable_skills": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "description": "Skill IDs to enable for this task. If empty or omitted, loads the skills the user has enabled in their account settings. Use [skill.list](https://open.manus.ai/docs/v2/skill.list) to retrieve available skill IDs."
                    },
                    "force_skills": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "description": "Skill IDs the agent must invoke during this task. Forced skills are automatically available even if not listed in enable_skills."
                    },
                    "task_references": {
                        "type": "array",
                        "items": {
                            "type": "string",
                            "pattern": "^[A-Za-z0-9]{22}$"
                        },
                        "maxItems": 20,
                        "description": "Task IDs of existing tasks to reference from this message. The agent can then browse the conversation and files of those tasks on demand. Each entry must be a bare 22-character alphanumeric task ID — a full task URL, or a shortcut value such as `agent-default-main_task`, is rejected. Validated at request time: any malformed ID, or more than 20 entries, fails with `InvalidArgument` (HTTP 400). The 20-item cap is applied to the array as sent, before duplicates are removed, so 21 entries fail even if some repeat. References accumulate: IDs passed to [task.sendMessage](https://open.manus.ai/docs/v2/task.sendMessage) are added to the references the task already has, and an empty array does not clear them. Duplicates are dropped, and so is a self-reference on task.sendMessage. Beyond those request-time checks, access is enforced when the agent reads a reference, not when you send it: a task you cannot open is reported to the agent as not found and your request still succeeds. One exception: an OAuth App that holds only the create_task scope can reference only tasks it created itself, and gets 404 at request time otherwise."
                    }
                },
                "required": [
                    "content"
                ]
            },
...
            "ErrorResponse": {
                "type": "object",
                "description": "Standard error response format returned when a request fails.",
                "properties": {
                    "ok": {
                        "type": "boolean",
                        "example": false,
                        "description": "Always false for error responses."
                    },
                    "request_id": {
                        "type": "string",
                        "description": "Unique identifier for this API request, useful for debugging with support."
                    },
                    "error": {
                        "type": "object",
                        "description": "Error details.",
                        "properties": {
                            "code": {
                                "type": "string",
                                "description": "Machine-readable error code (e.g., \"invalid_argument\", \"not_found\", \"permission_denied\", \"rate_limited\")."
                            },
                            "message": {
                                "type": "string",
                                "description": "Human-readable error description explaining what went wrong."
                            }
                        }
                    }
                }
            }
        }
    }
}